% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor_active_id}
\alias{editor_active_id}
\title{Active document id}
\usage{
editor_active_id(allow_console = TRUE)
}
\arguments{
\item{allow_console}{If \code{TRUE}, the console is a valid "active"
document; otherwise only Source pane documents qualify.}
}
\value{
A list with \code{id} set to the active id, or \code{NULL} when
there is no active document.
}
\description{
Wraps \code{\link[rstudioapi:documentId]{rstudioapi::documentId()}}.
}
