% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor_insert}
\alias{editor_insert}
\title{Insert text at the cursor (or replace the current selection)}
\usage{
editor_insert(text, id = NULL)
}
\arguments{
\item{text}{Text to insert.}

\item{id}{Document id. When \code{NULL}, the active document is targeted.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:insertText]{rstudioapi::insertText()}}.
}
