% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor_save_all}
\alias{editor_save_all}
\title{Save every open document}
\usage{
editor_save_all()
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{.rs.api.documentSaveAll()}.
}
