% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor_select_range}
\alias{editor_select_range}
\title{Set the selection in a document}
\usage{
editor_select_range(ranges, id = NULL)
}
\arguments{
\item{ranges}{A \code{document_range} (or list of them) describing the
new selection.}

\item{id}{Document id. When \code{NULL}, the active document is targeted.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:setSelectionRanges]{rstudioapi::setSelectionRanges()}}.
}
