% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor_set_cursor}
\alias{editor_set_cursor}
\title{Move the cursor in a document}
\usage{
editor_set_cursor(position, id = NULL)
}
\arguments{
\item{position}{A \code{document_position} (or a length-2 integer vector
\code{c(row, column)}).}

\item{id}{Document id. When \code{NULL}, the active document is targeted.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:setCursorPosition]{rstudioapi::setCursorPosition()}}.
}
