% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_add_output}
\alias{job_add_output}
\title{Append output to a job's log}
\usage{
job_add_output(job, output, error = FALSE)
}
\arguments{
\item{job}{Job id.}

\item{output}{Output text to append.}

\item{error}{If \code{TRUE}, the output is styled as error in the pane.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Append output to a job's log
}
