% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_add_progress}
\alias{job_add_progress}
\title{Bump a job's progress by \code{units} (incremental)}
\usage{
job_add_progress(job, units)
}
\arguments{
\item{job}{Job id.}

\item{units}{Increment in the units established at creation time.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Bump a job's progress by \code{units} (incremental)
}
