% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_is_active}
\alias{job_is_active}
\title{Are we currently running inside a job's R process?}
\usage{
job_is_active()
}
\value{
A list with \code{is_job} (logical).
}
\description{
Wraps \code{\link[rstudioapi:isJob]{rstudioapi::isJob()}}. Returns \code{TRUE} only when called from
inside a job script.
}
