% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_remove}
\alias{job_remove}
\title{Remove a job from the Jobs pane}
\usage{
job_remove(job)
}
\arguments{
\item{job}{Job id.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Remove a job from the Jobs pane
}
