% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_set_progress}
\alias{job_set_progress}
\title{Mark a job's progress}
\usage{
job_set_progress(job, units)
}
\arguments{
\item{job}{Job id.}

\item{units}{New cumulative progress, in the units established at
creation time.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Mark a job's progress
}
