% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_set_state}
\alias{job_set_state}
\title{Mark a job as running/stopped/finished}
\usage{
job_set_state(job, state)
}
\arguments{
\item{job}{Job id.}

\item{state}{One of \code{"idle"}, \code{"running"}, \code{"succeeded"}, \code{"cancelled"},
\code{"failed"}.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Mark a job as running/stopped/finished
}
