% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job_set_status}
\alias{job_set_status}
\title{Update the status message shown next to a job}
\usage{
job_set_status(job, status)
}
\arguments{
\item{job}{Job id.}

\item{status}{New status message.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Update the status message shown next to a job
}
