% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pane_files}
\alias{pane_files}
\title{Navigate the Files pane to a directory}
\usage{
pane_files(path)
}
\arguments{
\item{path}{Absolute directory path.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:filesPaneNavigate]{rstudioapi::filesPaneNavigate()}}.
}
