% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pane_preview_rd}
\alias{pane_preview_rd}
\title{Render an .Rd file in the Help pane}
\usage{
pane_preview_rd(path)
}
\arguments{
\item{path}{Absolute path to an \code{.Rd} file.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:previewRd]{rstudioapi::previewRd()}}.
}
