% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pane_viewer}
\alias{pane_viewer}
\title{Display a URL/path in the Viewer pane}
\usage{
pane_viewer(target)
}
\arguments{
\item{target}{File path or URL to render.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Thin wrapper around \code{\link[rstudioapi:viewer]{rstudioapi::viewer()}} for the common case of
pointing the Viewer pane at a local HTML file or a URL.
}
