% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{pref_set_persistent}
\alias{pref_set_persistent}
\title{Write a persistent value}
\usage{
pref_set_persistent(name, value)
}
\arguments{
\item{name}{Preference name.}

\item{value}{Any R value (lists, scalars, vectors). Stored as-is.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:setPersistentValue]{rstudioapi::setPersistentValue()}}.
}
