% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{pref_write_rstudio}
\alias{pref_write_rstudio}
\title{Write a built-in RStudio preference}
\usage{
pref_write_rstudio(name, value)
}
\arguments{
\item{name}{Preference name.}

\item{value}{Any R value (lists, scalars, vectors). Stored as-is.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Write a built-in RStudio preference
}
