% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_current}
\alias{project_current}
\title{Active project path}
\usage{
project_current()
}
\value{
A character path or \code{NULL}.
}
\description{
Returns the path of the currently-open RStudio project, or \code{NULL}
when no project is active. Thin wrapper around
\code{\link[rstudioapi:getActiveProject]{rstudioapi::getActiveProject()}}.
}
