% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term_activate}
\alias{term_activate}
\title{Activate (give focus to) a terminal}
\usage{
term_activate(id)
}
\arguments{
\item{id}{Terminal handle.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:terminalActivate]{rstudioapi::terminalActivate()}}.
}
