% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term_clear}
\alias{term_clear}
\title{Clear a terminal buffer}
\usage{
term_clear(id)
}
\arguments{
\item{id}{Terminal handle.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:terminalClear]{rstudioapi::terminalClear()}}.
}
