% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term_exec}
\alias{term_exec}
\title{Send text to a terminal followed by a newline (i.e. "execute it")}
\usage{
term_exec(id, text)
}
\arguments{
\item{id}{Terminal handle.}

\item{text}{Text to send.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Convenience over \code{\link[=term_send]{term_send()}} that ensures a trailing newline.
}
