% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term_kill}
\alias{term_kill}
\title{Kill a terminal}
\usage{
term_kill(id)
}
\arguments{
\item{id}{Terminal handle.}
}
\value{
\code{NULL} invisibly. Side-effect only.
}
\description{
Wraps \code{\link[rstudioapi:terminalKill]{rstudioapi::terminalKill()}}. The process is terminated and
the terminal pane tab is removed.
}
