% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term_visible}
\alias{term_visible}
\title{Identifier of the currently-visible terminal}
\usage{
term_visible()
}
\value{
A list with \code{id} set to the visible terminal's handle, or
\code{NULL} if no terminal is open.
}
\description{
Wraps \code{\link[rstudioapi:terminalVisible]{rstudioapi::terminalVisible()}}.
}
